requirements.txt
aiohttp
discord
python-dotenv
